lost-and-found/requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
face-recognition>=1.3.0
dlib>=19.24.0
Pillow>=9.0.0
setuptools>=65.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
gunicorn>=21.0.0
